+++ example/android/app/src/main/AndroidManifest.xml.example
