+++ example/android/app/src/main/kotlin/com/example/google_map_place_picker/example/MainActivity.kt
@@ -1,12 +1,7 @@
 package com.example.google_map_place_picker.example
 
-import androidx.annotation.NonNull;
 import io.flutter.embedding.android.FlutterActivity
-import io.flutter.embedding.engine.FlutterEngine
-import io.flutter.plugins.GeneratedPluginRegistrant
 
 class MainActivity: FlutterActivity() {
-    override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
-        GeneratedPluginRegistrant.registerWith(flutterEngine);
-    }
+
 }
