+++ lib/providers/place_provider.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:google_maps_place_picker_mb/src/models/pick_result.dart';
@@ -9,9 +9,6 @@ import 'package:google_maps_webservice/geocoding.dart';
 import 'package:google_maps_webservice/places.dart';
 import 'package:http/http.dart';
 import 'package:provider/provider.dart';
-import 'package:get/get.dart';
-
-import '../src/models/circle_area.dart';
 
 class PlaceProvider extends ChangeNotifier {
   PlaceProvider(
@@ -44,65 +41,42 @@ class PlaceProvider extends ChangeNotifier {
   bool isOnUpdateLocationCooldown = false;
   LocationAccuracy? desiredAccuracy;
   bool isAutoCompleteSearching = false;
-  CircleArea? searchRadiusArea;
 
-  LocationPlatformInterface.Location location =
-      new LocationPlatformInterface.Location();
-  LocationPlatformInterface.PermissionStatus permissionGranted =
-      LocationPlatformInterface.PermissionStatus.denied;
-  bool isLocationServiceEnabled = false;
+  Future<void> updateCurrentLocation() async {
+    bool serviceEnabled;
+    LocationPermission permission;
 
-  Future<void> updateCurrentLocation(bool forceAndroidLocationManager) async {
-    isLocationServiceEnabled = await location.serviceEnabled();
-    if (!isLocationServiceEnabled) {
-      // isLocationServiceEnabled = await location.requestService();
-      if (!isLocationServiceEnabled) {
-        // Get.dialog(
-        //   Padding(
-        //     padding: const EdgeInsets.all(10.0),
-        //     child: Column(
-        //       mainAxisAlignment: MainAxisAlignment.center,
-        //       crossAxisAlignment: CrossAxisAlignment.center,
-        //       children: [
-        //         const Icon(Icons.location_disabled),
-        //         const SizedBox(
-        //           height: 10,
-        //         ),
-        //         const Text(
-        //             "Your location service seems to be deisabled. Please enable your Location service and Click/Tap the Continue button."),
-        //         const SizedBox(
-        //           height: 20,
-        //         ),
-        //         ElevatedButton(
-        //           onPressed: () async {
-        //             await requestPermission().whenComplete(() => Get.back());
-        //           },
-        //           child: const Text("Continue"),
-        //         ),
-        //       ],
-        //     ),
-        //   ),
-        // );
-        // return;
+    // Test if location services are enabled.
+    serviceEnabled = await Geolocator.isLocationServiceEnabled();
+    if (!serviceEnabled) {
+      // Location services are not enabled don't continue
+      // accessing the position and request users of the
+      // App to enable the location services.
+      return Future.error('Location services are disabled.');
+    }
+
+    permission = await Geolocator.checkPermission();
+    if (permission == LocationPermission.denied) {
+      permission = await Geolocator.requestPermission();
+      if (permission == LocationPermission.denied) {
+        // Permissions are denied, next time you could try
+        // requesting permissions again (this is also where
+        // Android's shouldShowRequestPermissionRationale
+        // returned true. According to Android guidelines
+        // your App should show an explanatory UI now.
+        return Future.error('Location permissions are denied');
       }
     }
 
-    await requestPermission();
-    notifyListeners();
-  }
-
-  Future<void> requestPermission() async {
-    permissionGranted = await location.hasPermission();
-    try {
-      permissionGranted = await location.requestPermission();
-      if (permissionGranted ==
-          LocationPlatformInterface.PermissionStatus.granted) {
-        currentPosition = await Geolocator.getCurrentPosition(
-            desiredAccuracy: desiredAccuracy ?? LocationAccuracy.best);
-      } else {
-        currentPosition = null;
-      }
+    if (permission == LocationPermission.deniedForever) {
+      // Permissions are denied forever, handle appropriately.
+      return Future.error(
+          'Location permissions are permanently denied, we cannot request permissions.');
     }
+
+    _currentPosition = await Geolocator.getCurrentPosition(
+      desiredAccuracy: desiredAccuracy ?? LocationAccuracy.best,
+    );
   }
 
   Position? _currentPosition;
@@ -135,10 +109,6 @@ class PlaceProvider extends ChangeNotifier {
   PickResult? get selectedPlace => _selectedPlace;
   set selectedPlace(PickResult? result) {
     _selectedPlace = result;
-
-    if (SearchRadius > 0)
-      searchRadiusArea = CircleArea(
-          center: _currentCameraPosition!.target, radius: SearchRadius);
     notifyListeners();
   }
 
@@ -167,13 +137,6 @@ class PlaceProvider extends ChangeNotifier {
   bool get isSearchBarFocused => _isSeachBarFocused;
   set isSearchBarFocused(bool focused) {
     _isSeachBarFocused = focused;
-    notifyListeners();
-  }
-
-  double _searchRadius = 0;
-  double get SearchRadius => _searchRadius;
-  set Searchradius(double rad) {
-    _searchRadius = rad;
     notifyListeners();
   }
