+++ pubspec.yaml
@@ -11,7 +11,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  geolocator: ^9.0.1
+  geolocator: ^9.0.2
   google_api_headers: ^1.5.0+1
   google_maps_flutter: ^2.1.10
   google_maps_webservice: ^0.0.20-nullsafety.5
